+++ Forge.Forms/src/Forge.Forms.Demo/Models/Login.cs
@@ -1,4 +1,5 @@
-﻿using System.ComponentModel;
+﻿using System;
+using System.ComponentModel;
 using System.Runtime.CompilerServices;
 using Forge.Forms.Annotations;
 using MaterialDesignThemes.Wpf;
@@ -7,7 +8,7 @@ namespace Forge.Forms.Demo.Models
 {
     [Title("Log in to continue")]
     [Action("cancel", "CANCEL", IsCancel = true, ClosesDialog = true)]
-    [Action("login", "LOG IN", IsLoading = "{Binding Loading}",
+    [Action("{Binding Readable}", "LOG IN {Binding Readable}", IsLoading = "{Binding Loading}",
         IsDefault = true, ClosesDialog = true, Validates = true)]
     public class Login : IActionHandler, INotifyPropertyChanged
     {
@@ -28,6 +29,8 @@ namespace Forge.Forms.Demo.Models
                 OnPropertyChanged();
             }
         }
+
+        public object Readable => DateTime.Now;
 
         // Or be dynamically assigned...
         [Field(Icon = "{Property PasswordIcon}")]