+++ Forge.Forms/src/Forge.Forms/Annotations/DirectContentAttribute.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Forge.Forms.Annotations
+{
+    /// <summary>
+    /// Specifies that content rendering should be handled by WPF.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Property)]
+    public class DirectContentAttribute : Attribute
+    {
+    }
+}