+++ Forge.Forms/src/Forge.Forms/DynamicExpressions/DataBinding.cs
@@ -9,22 +9,35 @@ namespace Forge.Forms.DynamicExpressions
     public sealed class DataBinding : Resource
     {
         public DataBinding(string propertyPath, BindingOptions bindingOptions)
-            : this(propertyPath, bindingOptions, null, null)
+            : this(propertyPath, bindingOptions, null, null, false)
         {
         }
 
         public DataBinding(string propertyPath, BindingOptions bindingOptions, List<IValidatorProvider> validationRules)
-            : this(propertyPath, bindingOptions, validationRules, null)
+            : this(propertyPath, bindingOptions, validationRules, null, false)
+        {
+        }
+
+        public DataBinding(string propertyPath, BindingOptions bindingOptions, 
+            List<IValidatorProvider> validationRules, bool oneWay)
+            : this(propertyPath, bindingOptions, validationRules, null, oneWay)
+        {
+        }
+
+        public DataBinding(string propertyPath, BindingOptions bindingOptions, 
+            List<IValidatorProvider> validationRules, string valueConverter)
+            : this(propertyPath, bindingOptions, validationRules, valueConverter, false)
         {
         }
 
         public DataBinding(string propertyPath, BindingOptions bindingOptions, List<IValidatorProvider> validationRules,
-            string valueConverter)
+            string valueConverter, bool oneWay)
             : base(valueConverter)
         {
             PropertyPath = propertyPath;
             BindingOptions = bindingOptions ?? throw new ArgumentNullException(nameof(bindingOptions));
             ValidationRules = validationRules ?? new List<IValidatorProvider>();
+            OneWay = oneWay;
         }
 
         public string PropertyPath { get; }
@@ -33,11 +46,18 @@ namespace Forge.Forms.DynamicExpressions
 
         public List<IValidatorProvider> ValidationRules { get; }
 
+        public bool OneWay { get; }
+
         public override bool IsDynamic => true;
 
         public override BindingBase ProvideBinding(IResourceContext context)
         {
             var binding = context.CreateModelBinding(PropertyPath);
+            if (OneWay)
+            {
+                binding.Mode = BindingMode.OneWay;
+            }
+
             binding.Converter = GetValueConverter(context);
             BindingOptions.Apply(binding);
             var pipe = new ValidationPipe();